+++ index.php
@@ -9,6 +9,10 @@
     <?php
         echo GenHead::GenCSS(GlobalVars::$CSS);
         echo GenHead::GenScripts(GlobalVars::$Scripts);
+        $OpenGraph = array(
+            
+        );
+        echo GenHead::GenSocialCards(GlobalVars::$Twitter, $openGraph);
     ?>
     <link rel="shortcut icon" href="/favicon.ico" type="image/x-icon">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">