+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 Versioning numbers defined as MAJOR.MINOR.PATCH
 
+## [0.7.0] - 2018-03-03
+### Added
+- Automatic twitter metadate
+- GenHead class 'GenSocialCards' function
+- GlobalVars class 'Twitter' Static Variable
+
 ## [0.6.1] - 2018-03-03
 ### Fixed
 - GenHead.class.php, undefined index 'Media', line 23
@@ -12,6 +18,7 @@ Versioning numbers defined as MAJOR.MINOR.PATCH
 ## [0.6.0] - 2018-03-03
 ### Added
 - GenHead class 'GenScripts' function
+- GlobalVars class 'Scripts' Static Variable
 
 ## [0.5.1] - 2018-03-03
 ### Fixed